requirements.txt
streamlit
numpy
opencv-python-headless
pillow
git+https://github.com/facebookresearch/segment-anything.git
torch
torchvision
streamlit-image-coordinates
